Further merging adjustements
